samples/source-rss/pyproject.toml
[tool.poetry]
name = "st-source-rss"
version = "0.1.0"
description = "Storytrends - rss source"
authors = ["Francois Vermaut <fvermaut@gmail.com>"]
readme = "README.md"
packages = [{include = "st_module"}]

[tool.poetry.dependencies]
python = "3.10.13"
pydantic = "2.6.3"
requests = "2.31.0"
pandas = "2.2.1"
python-dotenv = "1.0.1"
supabase = "^2.4.5"
newspaper3k = "^0.2.8"
lxml-html-clean = "^0.1.1"
feedparser = "^6.0.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
